+++ library/src/utils/entriesFromList/entriesFromList.test-d.ts
@@ -1,12 +1,23 @@
 import { describe, expectTypeOf, test } from 'vitest';
-import { string } from '../../schemas/index.ts';
+import { arrayAsync, string } from '../../schemas/index.ts';
 import { entriesFromList } from './entriesFromList.ts';
 
 describe('entriesFromList', () => {
-  test('should return object entries', () => {
-    const schema = string();
-    expectTypeOf(entriesFromList(['foo', 'bar', 'baz'], schema)).toEqualTypeOf<
-      Record<'foo' | 'bar' | 'baz', typeof schema>
-    >();
+  describe('should return object entries', () => {
+    const symbol = Symbol();
+
+    test('for sync schemas', () => {
+      const schema = string();
+      expectTypeOf(entriesFromList(['foo', 123, symbol], schema)).toEqualTypeOf<
+        Record<'foo' | 123 | typeof symbol, typeof schema>
+      >();
+    });
+
+    test('for async schemas', () => {
+      const schema = arrayAsync(string());
+      expectTypeOf(entriesFromList(['foo', 123, symbol], schema)).toEqualTypeOf<
+        Record<'foo' | 123 | typeof symbol, typeof schema>
+      >();
+    });
   });
 });